started animations

+++ TodoList/Views/ListView.swift
@@ -12,18 +12,22 @@ struct ListView: View {
     @EnvironmentObject var listViewModel: ListViewModel
         
     var body: some View {
-        VStack{
-            List{
-                ForEach(listViewModel.items) {item in
-                    ListRowView(item: item)
-                        .onTapGesture {
-                            withAnimation(.linear){
-                                listViewModel.updateItem(item: item)
+        ZStack{
+            if listViewModel.items.isEmpty{
+                Text("No Items")
+            } else {
+                List{
+                    ForEach(listViewModel.items) {item in
+                        ListRowView(item: item)
+                            .onTapGesture {
+                                withAnimation(.linear){
+                                    listViewModel.updateItem(item: item)
+                                }
                             }
-                        }
+                    }
+                    .onDelete(perform: listViewModel.deleteItem)
+                    .onMove(perform: listViewModel.moveItem)
                 }
-                .onDelete(perform: listViewModel.deleteItem)
-                .onMove(perform: listViewModel.moveItem)
             }
         }
         .listStyle(PlainListStyle())
